update lesson 4

+++ TodoList/TodoList/Views/ListView.swift
@@ -12,17 +12,23 @@ struct ListView: View {
     @EnvironmentObject private var listViewModel:ListViewModel
     
     var body: some View {
-        List {
-            ForEach(listViewModel.items) { item in
-                ListRowView(item: item)
-                    .onTapGesture {
-                        withAnimation(.linear) {
-                            listViewModel.update(item: item)
-                        }
+        ZStack {
+            if listViewModel.items.isEmpty {
+                Text("No data!")
+            }else {
+                List {
+                    ForEach(listViewModel.items) { item in
+                        ListRowView(item: item)
+                            .onTapGesture {
+                                withAnimation(.linear) {
+                                    listViewModel.update(item: item)
+                                }
+                            }
                     }
+                    .onDelete(perform: listViewModel.deleteItem)
+                    .onMove(perform: listViewModel.moveItem)
+                }
             }
-            .onDelete(perform: listViewModel.deleteItem)
-            .onMove(perform: listViewModel.moveItem)
         }
         .navigationTitle("Todo List 📝")
         .toolbar {
